+++ src/styles/objects/buttons.js
@@ -1,19 +1,17 @@
 import styled from 'styled-components';
 
 export const Button = styled.a`
-  & {
-    display: block;
-    padding: 1rem var(--gap);
-    color: ${props => props.color};
-    background: ${props => props.background};
-    box-shadow: 0px 1px 10px rgba(0, 0, 0, 0.2), 0px 4px 5px rgba(0, 0, 0, 0.12),
-      0px 2px 4px rgba(0, 0, 0, 0.14);
-    border-radius: 4px;
-    max-width: 12rem;
-    margin: 0 auto;
-    text-align: center;
-    transition: filter 0.2s;
-  }
+  display: block;
+  padding: 1rem var(--gap);
+  color: ${props => props.color};
+  background: ${props => props.background};
+  box-shadow: 0px 1px 10px rgba(0, 0, 0, 0.2), 0px 4px 5px rgba(0, 0, 0, 0.12),
+    0px 2px 4px rgba(0, 0, 0, 0.14);
+  border-radius: 4px;
+  max-width: 12rem;
+  margin: 0 auto;
+  text-align: center;
+  transition: filter 0.2s;
 
   &:hover {
     color: ${props => props.color};
